+++ Conditions/ConditionFunctions.cs
@@ -264,105 +264,4 @@ namespace Conditions
     // Class holding parameters and can call functions.  inherit from this, and override find function to add on functions
     // done this way for historical reasons instead of having a set of ptrs to classes handling functions.
 
-    public class ConditionFunctionHandlers
-    {
-        public class Parameter
-        {
-            public string value;
-            public bool isstring;
-        };
-
-        public List<Parameter> paras;
-
-        protected ConditionFunctions caller;
-        protected ConditionVariables vars;
-        protected ConditionPersistentData persistentdata;
-        protected int recdepth;
-
-        protected delegate bool func(out string output);
-
-        protected class FuncEntry
-        {
-            public string fname;
-            public int numberparasmin;
-            public int numberparasmax;
-            public int checkvarmap;             // if not a string, check macro and fail if does not exist. b0 = para1, etc.
-            public int allowstringmap;          // allow a string map. b0 = para1, etc.
-
-            public FuncEntry(func f, int min, int max, int checkmacromapx, int allowstringmapx )
-            {
-                fname = f.Method.Name;
-                numberparasmin = min; numberparasmax = max;
-                checkvarmap = checkmacromapx; allowstringmap = allowstringmapx;
-            }
-        }
-
-        protected const int NoMacros = 0;           // ID for checkvarmap above
-        protected const int FirstMacro = 1;
-        protected const int SecondMacro = 2;
-        protected const int FirstSecondMacro = 3;
-        protected const int FirstSecondThirdMacro = 7;
-        protected const int SecondOnMacro = 0xffffffe;
-        protected const int AllMacros = 0xfffffff;
-
-        protected const int NoStrings = 0;          // ID for allowstringmap above
-        protected const int FirstString = 1;
-        protected const int SecondString = 2;
-        protected const int SecondOnStrings = 0xffffffe;
-        protected const int AllStrings = 0xfffffff;
-
-        protected static System.Globalization.CultureInfo ct = System.Globalization.CultureInfo.InvariantCulture;
-        protected static Random rnd = new Random();
-
-        protected virtual FuncEntry FindFunction(string name) { return null; }
-
-        public ConditionFunctionHandlers(ConditionFunctions c, ConditionVariables v, ConditionPersistentData h, int recd)
-        {
-            caller = c;
-            vars = v;
-            persistentdata = h;
-            recdepth = recd;
-            paras = new List<Parameter>();
-        }
-
-        public bool RunFunction(string fname, out string output)
-        {
-            FuncEntry fe = FindFunction(fname.ToLower());       // function names are case insensitive
-
-            if (fe != null)
-            {
-                if (paras.Count < fe.numberparasmin)
-                    output = "Too few parameters";
-                else if (paras.Count > fe.numberparasmax)
-                    output = "Too many parameters";
-                else
-                {
-                    for (int i = 0; i < paras.Count; i++)
-                    {
-                        if (paras[i].isstring)
-                        {
-                            if (((fe.allowstringmap >> i) & 1) == 0)
-                            {
-                                output = "Strings are not allowed in parameter " + (i + 1).ToString(ct);
-                                return false;
-                            }
-                        }
-                        else if (((fe.checkvarmap >> i) & 1) == 1 && !vars.Exists(paras[i].value))
-                        {
-                            output = "Variable " + paras[i].value + " does not exist in parameter " + (i + 1).ToString(ct);
-                            return false;
-                        }
-                    }
-
-                    System.Reflection.MethodInfo mi = GetType().GetMethod(fe.fname, System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    func fptr = (func)Delegate.CreateDelegate(typeof(func), this, mi);      // need a delegate which is attached to this instance..
-                    return fptr(out output);
-                }
-            }
-            else
-                output = "Does not exist";
-
-            return false;
-        }
-    }
 }
